knowledge_graph/requirements.txt
neo4j>=5.18            # official Python driver
requests>=2.31         # Ollama client (reused from rag_classification)
boto3>=1.34            # only needed for LLM_PROVIDER=bedrock
python-dotenv>=1.0
